Fix SiteHeader not marking single-pages as 'selected' in the Navigation
When SiteHeader() creates the HTML it marks a Link in the Navigation menu as selected, when the `selectedSectionID` that was passed is the same as the id of the current Link.
This doesn't work for Single-Pages though, since they don't have a section by default in publish.
With this commit even single-pages can have their entry in the Navigation bar marked as 'selected'